Production ready schema TAKE 1!
