+++ src/data/chat-channel/index.js
@@ -1,0 +1,1 @@
+export { default as schema } from './schema';
