+++ src/data/chat-channel/schema.js
@@ -1,0 +1,14 @@
+import gql from 'graphql-tag'
+
+export default gql`
+    type ChatChannel {
+        name: String!
+        channelID: String!
+    }
+
+    interface ChatChannelNode {
+        chatChannel: ChatChannel
+    }
+
+    extend type LiveStream implements ChatChannelNode
+`
