+++ src/data/events/schema.js
@@ -13,6 +13,7 @@ export default gql`
   }
   
   extend type Campus {
-    events: [Event] @deprecated(reason: "Events are no longer organized by limiting them to a physical campus. Events are organized into categories. Please reference 'EventContentItem' instead")
+    events: [Event] 
+      @deprecated(reason: "Events are no longer organized by limiting them to a physical campus. Events are organized into categories. Please reference 'EventContentItem' instead")
   }
 `
